Add intermediate CA keypair
In preparation for generating cert chains, create a hard-coded
private key for use in intermediate CA certificates.  Also
create the corresponding fragment of ASCII ASN.1 to hold the
corresponding public key information so it can be included in
TBS certificates.  Modify the 2 existing example CA certs to
use this public key.