Improve clarity/correctness of BIT_STRINGs
The contents of a bit string starts with an octet indicating the
number of unused bits in the final octet; this needs to be manually
included.

Divide hex strings up to make this clear, e.g.
  `07` `ff80`
indicates that there are 7 unused bits in the final octet, so the
complete bitstring is b111111111.

Also remove enclosing BIT_STRING {} for UniqueIdentifier instances